FAB-16695 Enable split admin, cluster ports in IT
There was a bug where if the prometheus metrics provider was used, and
either the peer admin service port, or the orderer cluster port were set
to be different from the primary listen port, then metrics were double
registered, resulting in a peer crash.  This CR enables splitting these
ports in the integration tests, to confirm the bug has been eliminated,
and to detect other such regressions in the future.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>
Change-Id: I52d3721a51e80e1aefd15d59ca14bba9669bdd91