FAB-16695 Re-use metrics for comm GRPCServer
Today, the comm.NewGRPCServer code creates and registers metrics, given
a supplied metrics provider.  This is problematic because we should only
have one metrics provider, and re-registering a metric causes a panic.

Thus, if the admin service, or the Raft cluster service are put onto
different ports when Prometheus is enabled, the result is a crash.  This
CR fixes both peer and orderer initialization to re-use the metric
components across servers.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>
Change-Id: I80a33c1a3bd4cb25b08ca09297b098b11963dbba