FAB-16695 Make peerServer singleton a slice
The 'peerServer' is currently a singleton, which, when NewPeerServer is
called, is overridden.  However, if the admin server and the endorser
server are launched on different ports, then we will invoke
NewPeerServer multiple times, and overwriting the singleton.  This means
that when the ports are split, only one of the servers will get TLS cert
updates.  This simple change modifies the singleton to be a slice, so
that each server can get the updates.

Note, in master, this singleton is already gone, so this problem does
not exist.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>
Change-Id: I75e30d18d94846a9caabd00f65afe5a148b81948