[FAB-16651] Fix conn leak if certs renewed
When a peer rotates its certificate but preserves the endpoint,
the PKI-ID of the peer changes.

the membership module treats the old PKI-ID as a dead peer,
and asks the communication module to open a connection to it,
which it does, however - it associates the connection with the
*new* PKI-ID and not the old.

This is by design, to support connecting to bootstrap peers who's
PKI-IDs are unknown (and appear nil).

Since the old PKI-ID never appears alive - the membership module keeps
piling up connection creations that override the *new* PKI-ID in the mapping
and this eventually leads to a file descriptor overload.

This change set adds a feedback mechanism from the communication module
to the membership module that purges the old PKI-ID from memory in case
the PKI-ID has changed.

Now, the membership module does not try to re-connect to these old PKI-IDs,
because they are purged from memory.

Change-Id: I188b0e5a27d7f030e12b732ddc6efe557a19e014
Signed-off-by: yacovm <yacovm@il.ibm.com>