FAB-16806 address flaky test with async execution
The custodian Work function is started in a go routine. That routine
calls Build before attempting to close a channel that's used as a
completion signal.

There's a window between Build and the call to Notify where the test can
evaluate the completion channel closed by notify where the test can run.
That means the assertion should be async (Eventually) instead of
synchronous (Expect).

Change-Id: I3a043fa072476afc1fddd9ea758283af49c34ac6
Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>