FAB-16599 Add externalbuilder md integration tests
This CR does a few related things.

1) It modifies the existing integration test external builder to simply
refer to the chaincode by path, rather than bothering to package the
chaincode binary.  This eliminates the overloading of the Chaincode.Path
field which was being used both to carry the filesystem path of the
chaincode binary as well as the metadata JSON path info.

2) It modifies the existing chaincode db index tests using the new
lifecycle to use the externalbuilders for packaging.

3) It adds a bit of tracing for assembling the metadata, and
fixes a flaw in the header name discovered by the IT.

This CR still requires a bit of followup, notably, we should probably
add back a 'new lifecycle, conventional packaging, with db indices'
test.  Additionally, we need to add a 'legacy lifecycle, with external
builders, with chaincode indices'.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>
Change-Id: Ibab38f4ac8d6b9a1d7be8feeb3ae3f4ca0c79254