lifecycle int. helper should use endorser for query
The RunQueryInvokeQuery helper function for lifecycle integration tests
should use one of the peers that endorsed the chaincode invocation the
second time it queries the chaincode. Otherwise, a flake is possible since
other peers in the channel won't necessarily have committed the transaction
to their ledgers yet.

Also, for future debugging purposes, update the assertions to use WithOffset
to make it clear which call to the helper function resulted in a failure.

FAB-16878 #done

Change-Id: Ie1019c65d33fce7bfa58d9787be6d1e6dfbcbe69
Signed-off-by: Will Lahti <wtlahti@us.ibm.com>