Remove PROJECT_FILES from make process
The PROJECT_FILES macro is set to the list of all files in the git repo
that are not part of the integration or vendor folders. This list of
files is then used as the dependency list in a handful of make recipes.

As our makefile is largely a list of phony targets and source
dependencies are not really handled, remove PROJECT_FILES to simplify
the process and prevent consumers from assuming the file list is more
useful than it is.

FAB-16934 #done

Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>
Change-Id: Ic53d27e926cf5ec249ebf4758685b93ef9e1454a