FAB-15285 QSCC now rejects cc2cc invocations
This CR causes QSCC to reject invocations made in a cc2cc manner before
attempting to acquire ledger resources for the corresponding ledger.
This breaking change is required because any cc2cc invocation of QSCC is
inherently unsafe, and if made under load, has a high probability of
creating a deadlock as QSCC attempts to acquire a second read lock while
holding a first.  If the ledger attempts to acquire a write lock, then
the second read lock will block indefinitely as the write lock waits for
the initial read lock to be released.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>
Change-Id: I78f2e61aa6d30c20b34a23cff89b37b44851301c