UpdateOrdererConfigSession returns without waiting
For consistency with the rest of the *Session functions in the nwo
package, UpdateOrdererConfigSession will return the created command
session without waiting for it to complete.

To prevent the serialized block from getting deleted, the network root
directory was used as the parent of the temp directory for the update
block and the defer statement to remove the temp dir was removed from
the function.

Change-Id: I635aac0a0dc1d81df99f0b7a4e27e5a2e250a329
Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>