'release' output does not require META-INF prefix
When using an external builder, the `release` executable is provided
with the directory to populate with metadata information for the peer to
consume. Given it's defined role, it doesn't make sense to require an
additional META-INF subdirectory within it.

Update the metadata provider to implicitly prepend the META-INF prefix
when packaging the output and update the integration test implementation
to adhere to the contract.

FAB-16861 #done

Change-Id: Ia64bd6bb91a72fcf3a0e9a6a8ced249fd70bfd16
Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>