[FAB-15389] Fix private data dissemination
Endorsing peer was not honoring maxPeerCount for private data
dissemination due to the addition of the spraying logic for balancing
selected peers for disseminiation. There was a chance that peers chosen
from the remaining count up to maxPeerCount could overlap with the
already selected peers for dissemination from the spray, causing the
dissemination plan to be incomplete. This CR ensures that once a peer
has been selected to disseminate private data to, it can not be selected
again when computing the dissemination plan for a particular tx. Also
added logging around dissemination for debugging purposes.

Change-Id: Ib31c37b036363787718dba74e7e90a7764afe7fb
Signed-off-by: Danny Cao <dcao@us.ibm.com>