External launch metadata uses PEM encoding
The `chaincode.json` generated by the peer for use with external
builders and launchers uses strings to hold the PEM encoded certificates
and keys provided to the chaincode. This change prevents base64 encoding
of the (already) base64 PEM encoded materials.

In practical terms, we no longer need to use `base64 --decode` in the
run scripts.

FAB-17007 # done

Change-Id: I3ab3818ec5e82e86a2bb7975720384236163bd63
Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>