FAB-17170 externalbuilder pass metadata unmangled
Today, the external builder receives a structure representing the three
fields of the metadata necessary to invoke the old docker platform
builds.  It then takes these fields and reserializes them to the JSON
metadata document.

In fact, the externalbuilder has no need for these deserialized fields,
and should pass the JSON document through without modifying it or losing
any additional fields.  This CR causes the package retrieving code to
return both the JSON document and the underlying bytes.  The underlying
bytes are passed to the externalbuilder while the three field structure
is passed to the dockerbuilder.

These paths are really quite bifurcated and could probably be further
split, with different package retrieving calls, it didn't seem worth it
purely to address this bug.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>