Attempt to fix flaky integration ApproveForMyOrg
The lifecycle tests use the PeersWithChannel function to get a list of
peers to invoke ApproveChaincodeForMyOrg on.  That function then picks
the first peer from each org in this list.  Consequently, a random
subset of peers from orgs are picked.  Then sometimes, the peers will
have the implicit collection private data in their transient store, and
sometimes, they will not.

By making this function deterministic, we should ensure that these tests
consistently get Peer1 for their org, which is almost always the peer
which is selected for individual approvals.

This is untested (except soon to be in CI) but will hopefully reduce the
flakiness of these tests.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>