Remove CKA_PRIVATE from pubkeyT to broaden compatibility with HSMs
Some HSM does not support CKA_PRIVATE defined as 'false' for the pubkey
template. Removing this definition of the template eliminates the issue,
as the HSM assumes a default value for the created object.

There is another attribute that defines if a key object is the public or
private member of a pair, and this is the CKA_CLASS attr, that may have the
values of CKO_PUBLIC_KEY, CKO_PRIVATE_KEY or CKO_SECRET_KEY.

JIRA: https://jira.hyperledger.org/browse/FAB-17280

Tested with SoftHSMv2 and Dinamo Networks HSM.

Signed-off-by: Carlos Eduardo Matos Ellery <carlos.ellery@serpro.gov.br>