Use `go list -json` to build go cc dependencies
Prior to these changes, custom format templates were used to format
dependency information as JSON for go chaincode. This did not work on
Windows because the backlash path separators were not properly escaped.

Using the -json option (instead of the template) results in properly
escaped filenames in the structure so we abandon the custom format and
use the JSON output to generate our model.

Later, when we create the tar that's embedded in the chaincode, we
transform the path separators to slashes.

FAB-17288

Change-Id: If3e04f5fa4deacb734e94323c9872406f3f6406e
Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>