Remove t.Parallel from common/cluster tests
This change trades execution time for (hopefully) a bit more stability
in the cluster tests. We see that on constrained CI systems these tests
timeout. When they timeout, there are many, many go routines stuck in
t.Parallel waiting for tests to complete. These routines make it harder
to see where the true failures are.

So, let's try running each of these tests serially; by doing this we go
from local execution of about 8s to 18s. In the scope of tests that take
tens of minutes, this seems reasonable.

Change-Id: Icaea901bf03a6e15e6f9c32da2d7bedbdf7cdb38
Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>