[FAB-17308] Discovery supports private data blind writes
This change set adds an ability to hint that even though a peer is not
in a collection, but the collection was mentioned in the endorsement query,
the peer can be included in the endorsement response despite not being
in the collection.

This is needed to support private data "blind writes" where we want to
populate a pre-image and we need endorsements from peers that might not
be in the collection.

This commit:

- Re-imports protobuf to include the NoPrivateReads flag.
- Implements the required functionality in the discovery service
  and tests it with unit tests
- Adds a flag to the discover CLI
- Documents the new added flag
- Adds a check to the integration test

Change-Id: I75247aeda170e539216562c3b2baf5f7ef752b3a
Signed-off-by: yacovm <yacovm@il.ibm.com>