Static leader should not give up on retrieving blocks (#507)
When a peer is configured as a static org leader for gossip, it should
not give up on retrieving blocks in the event that the orderering
service goes down for an extended period of time (i.e. longer than the
peer.deliveryclient.reconnectTotalTimeThreshold). Otherwise, the peer
must be restarted before it will act as the leader again.

FAB-17327

Signed-off-by: Brett Logan <brett.t.logan@ibm.com>
Signed-off-by: Will Lahti <wtlahti@us.ibm.com>

Co-authored-by: Brett Logan <brett.t.logan@ibm.com>