Fail fast if 'jq' is missing
If 'jq' is missing from the PATH then the chaincode installation fails and it
is not crystal clear why it failed unless you inspect the logs with care.

This change set makes the integration tests fail fast when attempting to install
chaincode and 'jq' doesn't exist in the path.

It prints:

'jq' is needed to build chaincode but it wasn't find in the PATH

and terminates the test.

Change-Id: Ic730ec6b90343c1da06818b16fb291963cbc20ac
Signed-off-by: yacovm <yacovm@il.ibm.com>