Reduce execution time of platforms/golang tests
Packaging chaincode involves creating a compressed tar of the chaincode
source and its transitive dependencies. In many cases, this is hundreds
of files due to the dependency tree of grpc. It turns out that
compressing hundreds of files takes a while and doing it multiple times
in tests results in very long test execution times.

By disabling gzip compression, the tests (running with race detection
enabled) go from taking about 1m to about 10s.

This change creates an unexported package level var to hold the
compression level to use that is initialized to gzip.DefaultCompression.
An `init()` function was added to the test to explicitly set compression
to gzip.NoCompression.

Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>
Change-Id: Iaa454bf7873999aa484a4d328a6a7644a716012a