FAB-16297 remove pvtdataStore initiation with non-zero blkHt (#701)
Remove functions that set the initial state for the pvtdata
store when an existing block store has a blockchain but
the pvtdata store is empty. This situation is expected to
happen when a peer is upgraded directly from version 1.0
to a future version (i.e., 1.0+), and an existing blockchain
is present (which was created with version 1.0).  Under
this scenario, the pvtdata store is brought upto the
point as if it has processed existing blocks with no pvtdata.

As we don't expect/support a direct upgrade from
version 1.0 to version 2.1, we can remove code that takes
care of the above mentioned scenario..

Note that pvt data feature is available since
version 1.1 (experimental) and became stable in version 1.2

Signed-off-by: senthil <cendhu@gmail.com>