[FAB-17463] Allow peer to override implicit collection dissemination properties
Added the following config properties to peer's core.yaml for implicit collection
dissemination policy. When a peer endorses a proposal that writes to its own
implicit collection, these values override the default values.
- peer.gossip.pvtData.ImplicitCollectionDisseminationPolicy.requiredPeerCount
- peer.gossip.pvtData.ImplicitCollectionDisseminationPolicy.maxPeerCount

The properties are applicable to all channels the peer has joined. The implication
is that requiredPeerCount has to be smaller than the number of peers in a channel
that has the lowest numbers of peers from the organization.

Signed-off-by: Wenjian Qiao <wenjianq@gmail.com>