[FAB-16951] Alternative mechanisms to find pkcs11 key
This modification adds a parameter called AltID to the PKCS11 BCCSP module.
When this parameter is set, the generateECKey and findKeyPairFromSKI functions
will use this alternative ID as the CKA_ID and CKA_LABEL to store and use the
associated private key.
This code change is required in situations where the HSM does not allow
modification of the CKA_ID after creation.
This code change includes corresponding unit-test and has been tested against
both AWS CloudHSM and the SoftHSM.

Signed-off-by: Luc Desrosiers <ldesrosi@uk.ibm.com>