[FAB-17535] Fix intermittent test failure in implicit_coll_test
In implicit_coll_test.go, requiredPeerCount was originally set to 0,
so endorsement will return without waiting for the pvtdata to be
disseminated to other peer. As a result, the other peer may or may not
receive the disseminated pvtdata at the time of query (depending on
which is done first, commit+query vs. dissemination).

This PR changes requiredPeerCount to 1 in order to guarantee
that the other peer receives the disseminated pvtdata during endorsement.

Signed-off-by: Wenjian Qiao <wenjianq@gmail.com>