[FAB-15570] Make peer functions synchronous
The startup of peers was done an asynchronous
fashion which caused the system to slow down
when multiple peers were starting at once. This
caused the tests to often hit timeout limits.

This change modifies the startup to occur sequentially.
Timing show a reduction in actual start of 4x by reducing
deadlock in the system

Signed-off-by: Brett Logan <brett.t.logan@ibm.com>