Fix data race in TestBlockingEnqueue
In some tests we need to overwrite the mocks.Mock field
of the mock, however it could be that it is used concurrently
by other goroutines.

We use a lock to ensure mutual exclusion and memory cohesion,
and the mock methods lock before calling the mocked method.

The lock was only used upon demand and not consistently.

I made the lock to be always activated.

Change-Id: Iffd02f681b144d2456640922c4c5a6438109a40e
Signed-off-by: yacovm <yacovm@il.ibm.com>