[FAB-17624] Handle net.Listen error in gossip unit test
The function createDiscoveryInstanceThatGossipsWithInterceptors attempts
to create a net.Listener which binds on an address, and it can fail
if the port is in use.

The code currently just prints the error and continues, and later on
panics with a nil pointer.

I changed the code so that it now panics and doesn't continue the flow
of the test.

Change-Id: Ie55317a79d8fab03ab2f64d1e4a016cc416397ee
Signed-off-by: yacovm <yacovm@il.ibm.com>