Use WithOptions to register log hook in test
The test output was pointing out a problem:

```
2020-03-16 19:29:10.284 EDT [orderer.common.server] TestConfigureClusterListener -> DPAN 03c Ignored key without a value.  ignored=zap.optionFunc(0x4799510)
```

Turns out that someone was adding the option as a log field instead of
an option. Building on that bug, the order of arguments to assert.Subset
were reversed so the assertions were always making sure that an empty
slice was a subset of a slice of 0..n values. <sigh/>

Fix the things.

Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>