Call signal.Notify from main go routine
Prior to this change, signal handling was setup and executing in a go
routine that was spawned during server initialization. This could result
in `signal.Notify` getting called after the "readiness" message for the
server is issued. Signals sent to the process within this window can be
lost.

This change ensures that `signal.Notify` is called before "readiness"
messages are written to help ensure that integration tests that exercise
signal processing get consistent results.

FAB-17585 #done

Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>