Reduce TestDBRequestTimeout timeout value
Unit tests for a PR recently failed when a DB request with a very small
timeout seems to have completed successfully.

While a microsecond is an impossibly short period of time for an http
request, a nanosecond is a thousand times shorter. (It's also the
smallest time.Duration we can specify.)

This modifies the test timeout to a Nanosecond to make impossible things
even less likely.

FAB-17375

Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>