[FAB-17667] move couchdb util package to statecouchdb package
We have a CouchDB util package that exposes API to read from and
write to CouchDB. This package is used only by the statecouchdb and
the tests package (i.e., ledger level integration test).

In general, a util package is justified only when an API needs to be
used by so many other packages. For e.g., if we have 10 packages and
each of them need a common API, it is not good to duplicate the API
code and instead create a separate util package.

Hence, we move all files in the CouchDB util package to the statecouchdb
package itself. Note that there is a lot of opportunities to refactor
couchdb package but we will do that in a separate PR.

Given that we have moved couchdb util to the statecouchdb package,
we can have only one TestMain. It would be hacky to reuse
existing testVDBEnv for testing couchDB too. Hence, we introduce
another testEnv called testCouchDBEnv to test couchDB utils.

Signed-off-by: senthil <cendhu@gmail.com>