FAB-17677 mv txvalidation flag util to txvalidator package from ledger
The txvalidation util package helps in creating transaction validation
flags that would be assigned to the block metadata. This function is used
by the v14 and v20 validation as it is needed by these packages to record
invalid tx as per policy evaluation. This validation flags is also reused
in the mvcc validator to mark valid/invalid transactions. In addition to
these two packages, the txflags are referred in so many other packages.
Keeping the txflag in core/ledger/util is not correct. Hence, we move the
txvalidation util package to the internal/pkgs/txflags.

Signed-off-by: senthil <cendhu@gmail.com>