merge ledgerstorage with kvledger pkg
The ledgerstorage package encapsulates both pvtdata storage and
block storage. We needed this package as

 - we had to have 2-phase commit between pvtdata storage and block storage.
 - we had a recovery logic that was needed after a commit failure or
   peer crash between pvtdata commit and blockstore commit.
 - we treated pvtdata store as a pvt blockstore and blockstore as
   public block store.

After introducing the support for peer rollback, we removed the 2-phase
commit between requirement between the pvtdata storage and block storage.
As a result, the pvtdata storage can be ahead of blockstore. Hence, the
recovery logic was also remoted. Hence, the motivation for having
ledgerstorage package became feeble and hence, we remove it.

Signed-off-by: senthil <cendhu@gmail.com>