[FAB-17676] mv cache package to statecouchdb
We have introduced the cache implementation in the statedb package
as the initial plan was to integrate cache with both statecouchdb
and stateleveldb. However, we have decided to defer the cache
integration with stateleveldb as the expected throughput improvement
would be < 5%. This is because the goleveldb is an embedded DB
(i.e., no cost associated with the access) and it utilizes OS
page cache anyway to avoid the disk access. Further, the goleveldb
is not a bottleneck as of now.

Hence, we move the cache implementation to the statecouchdb package
itself. When there is a need to integrate cache with goleveldb,
we can move it back.

Signed-off-by: senthil <cendhu@gmail.com>