rm unnecessary import dependency from statecouchdb package
The statecouchdb package should not be dependent on too many
other packages. It should be self-sufficient. Currently, we
have imported ccprovider just for a struct. Further, the test
imports ccprovider to use a function. As these are not
justified, we remove it.

Given that we pass map[string][]byte to the function that
processes the index data, we don't need to create a tarFile.

Signed-off-by: senthil <cendhu@gmail.com>