Build tools as module dependencies (except dep)
This is a small step towards modules by adopting the "tools.go" idiom to
describe and version the tool dependencies for the project without
bringing modules as a whole to fabric.

The tools.go file contains unnamed imports of the main packages of the
tools we use. The file itself has a build tag constraint to prevent its
presence from impacting most go tooling.

When installing the tools, instead of relying on vendored packages in
Fabric, the install is done in module mode within the tools context of
the tools module such that it picks up the versions of dependencies
pinned in the module.

The tools module dependencies are not currently vendored. While we can
(and maybe will) do this, the code supporting the tools is not linked
into the binaries produce by Fabric so we can still build without them.
Outside of linting and testing that's done in CI, the remaining tools
are related to code generation by developers.

As Fabric is still not a module, the tools will be installed to
${GOPATH}/bin as they currently are.

Since tools dependencies are getting moved to the tools module, they are
being removed from Gopkg.toml.

Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>