mv version package to core/ledger/internal/pkg
To ensure serializable isolation level for transaction, we have
implemented optimistic concurrency control (OCC) in v1.0. This
OCC attach version to each state update and uses it to detect
serilizable violations. As this package is core to ledger and
non-ledger packages do not need it, we move version package to
core/ledger/internal/pkg

Signed-off-by: senthil <cendhu@gmail.com>